minor improvements
shorten sys.setrecursionlimit before tests intended to overflow the
recursion limit (because on some platforms we seem to run out of
system stack first, which is Bad).

Do not bother to py_DECREF objects that were captured by stable pointers
if we are tearing down the module. (Seems that one can sometimes
block forever waiting for the GIL or something like that.)