Add tests for GIL and signal fancy behavior (#2)
Add tests for GIL and signal fancy behavior

There were previously no tests for the fancy behaviors where we (optionally) release the GIL while running Haskell code, and where we arrange that signals can interrupt the python program even when it is in Haskell code. Add tests for these things (the signal code is only tested on POSIX, because it seems to be very hard on windows to simulate sending control-C to a process).