update for changes upstream (#16)
* update for changes upstream

The Python abstract base classes have moved to collections.abc, and some error messages that we see in doctests were updated in Python 3.10.

Data.Text is beginning to switch over to a UTF8 internal rep. Support this. In service of that, when working with earlier versions of Data.Text, we now do some UTF8->UTF16 conversions on the Haskell side.

Also bump a stackage referrer.